--- oldjs/ga.js
@@ -1,15 +1,0 @@
-// Standard Google Universal Analytics code
-
-(function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
-
-    (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
-    
-    m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
-    })(window,document,'script','https://www.google-analytics.com/analytics.js','ga'); 
-
-ga("create", "UA-177701600-1", "auto");
-ga("set", "checkProtocolTask", function () {}); // Removes failing protocol check. @see: http://stackoverflow.com/a/22152353/1958200
-ga("require", "displayfeatures");
-ga("send", "pageview", window.location.href);   
-
-
