+++ src/app/utils/options.js
@@ -1,5 +1,5 @@
 let atag = document.getElementById("pmode");
-atag.onclick = () => pmode();
+atag.addEventListener("click", pmode);
 
 function pmode() {
   chrome.storage.sync.get(["pmode"], function (result) {
@@ -8,8 +8,6 @@ function pmode() {
 }
 
 chrome.storage.onChanged.addListener(function (result, storageArea) {
-  // console.log(result);
-  // console.log(storageArea);
   let mainView = document.getElementsByTagName("main");
   if (result["pmode"]) {
     if (result["pmode"].newValue === true) {
