+++ src/index.html
@@ -12,7 +12,6 @@
     />
     <link rel="shortcut icon" type='image/png' href="./assets/polus_tab_icon.png" />
   </head>
-
   <body>
     <ul id="nav">
       <li id="app-info-box">
@@ -22,7 +21,7 @@
             <span class='app-sub'>Polus</span>
           </div>
           <div id="app-items">
-            <div id="settings">Settings</div>
+            <!-- <div id="settings">Settings</div> -->
             <div id="site">
               <a
                 target="_blank"
@@ -77,23 +76,14 @@
         <button type="button" class="view-btn">Week</button>
         <button type="button" class="view-btn">Month</button>
       </div>
-
       <section id="daily"></section>
       <section id="week"></section>
       <section id="month"></section>
     </main>
 
-
-    <!-- <script src="https://www.gstatic.com/firebasejs/8.0.2/firebase-app.js"></script> -->
-    <!-- <script src="https://www.gstatic.com/firebasejs/8.0.2/firebase-auth.js"></script> -->
-    <!-- <script type="text/javascript" src="../js/auth.js"></script> -->
-    <!-- <script type="text/javascript" src="../js/view.js"></script>
-    <script type="text/javascript" src="../js/clock.js"></script>
-    <script type="text/javascript" src="../js/helper.js"></script>
-    <script type="text/javascript" src="../js/day.js"></script>
-    <script type="text/javascript" src="../js/week.js"></script>
-    <script type="text/javascript" src="../js/month.js"></script>
-    <script type='text/javascript' src='../js/options.js'></script>
-    <script type="text/javascript" src="../js/start.js"></script> -->
+    <!-- <script src="https://www.gstatic.com/firebasejs/8.0.2/firebase-app.js"></script>
+    <script src="https://www.gstatic.com/firebasejs/8.0.2/firebase-auth.js"></script>
+    <script type="text/javascript" src="../js/auth.js"></script>
+    <script type="text/javascript" src="../js/view.js"></script> -->
   </body>
 </html>
