+++ webpack.common.js
@@ -1,15 +1,7 @@
 module.exports = {
   entry: {
-    main: "./src/index.js",
+    index: "./src/index.js",
     popup: "./src/popup.js",
     background: "./src/background.js",
-  },
-  module: {
-    rules: [
-      {
-        test: /\.html$/i,
-        loader: "html-loader",
-      },
-    ],
   },
 };
