--- webpack.config.js
@@ -1,87 +1,0 @@
-const path = require("path");
-const HtmlWebpackPlugin = require("html-webpack-plugin");
-const MiniCssExtractPlugin = require("mini-css-extract-plugin"); // extract css
-const { CleanWebpackPlugin } = require("clean-webpack-plugin");
-const CopyPlugin = require("copy-webpack-plugin"); // used to copy files of any time to dist
-
-function makeHash(string) {
-  return string;
-}
-
-function modify(buffer, path) {
-  let manifest = JSON.parse(buffer.toString());
-  // console.log(manifest);
-
-  // make any modifications that you would like
-
-  // Pretty print to JSON with 2 spaces
-  manifest_JSON = JSON.stringify(manifest, null, 2);
-  return manifest_JSON;
-}
-
-module.exports = (env) => {
-  return {
-    mode: "production",
-    entry: {
-      index: "./src/index.js",
-      popup: "./src/popup.js",
-      background: "./src/background.js",
-    },
-
-    output: {
-      filename: "[name].js",
-      path: path.resolve(__dirname, "dist"),
-    },
-
-    module: {
-      rules: [
-        {
-          test: /\.scss$/,
-          use: [
-            MiniCssExtractPlugin.loader, // 3. Inject styles into DOM
-            "css-loader", // 2. Turns css into commonjs
-            "sass-loader", // 1. Turns sass to css
-          ],
-        },
-      ],
-    },
-
-    plugins: [
-      new CopyPlugin({
-        patterns: [
-          {
-            context: "src/",
-            from: "./assets/",
-            to: path.resolve(__dirname, "dist/assets"),
-          },
-        ],
-      }),
-      new CopyPlugin({
-        patterns: [
-          {
-            from: makeHash("./src/manifest.json"),
-            to: "./manifest.json",
-            transform(content, path) {
-              return modify(content, path);
-            },
-          },
-        ],
-      }),
-
-      new MiniCssExtractPlugin({ filename: "[name].[fullhash].css" }),
-      new HtmlWebpackPlugin({
-        template: "./src/index.html",
-        chunks: ["index"],
-        inject: true,
-        filename: "index.html",
-      }),
-      new HtmlWebpackPlugin({
-        template: "./src/popup.html",
-        chunks: ["popup"],
-        filename: "popup.html",
-        inject: true,
-      }),
-      new CleanWebpackPlugin(),
-    ],
-  };
-};
