+++ webpack.dev.js
@@ -1,16 +1,25 @@
 const path = require("path");
 const { merge } = require("webpack-merge");
-const MiniCssExtractPlugin = require("mini-css-extract-plugin"); // extract css
 const common = require("./webpack.common");
+
+const HtmlWebpackPlugin = require("html-webpack-plugin");
+const MiniCssExtractPlugin = require("mini-css-extract-plugin"); // extract css
+const { CleanWebpackPlugin } = require("clean-webpack-plugin");
 const CopyPlugin = require("copy-webpack-plugin"); // used to copy files of any time to dist
 
 module.exports = merge(common, {
   mode: "development",
 
+  devServer: {
+    contentBase: path.join(__dirname, "dist"),
+    compress: true,
+    port: 9000,
+  },
   output: {
     filename: "[name].js",
     path: path.resolve(__dirname, "dist"),
   },
+
   module: {
     rules: [
       {
@@ -23,14 +32,10 @@ module.exports = merge(common, {
       },
     ],
   },
+
   plugins: [
     new CopyPlugin({
       patterns: [
-        {
-          context: "src/",
-          from: "*.json",
-          to: path.resolve(__dirname, "dist"),
-        },
         {
           context: "src/",
           from: "./assets/",
@@ -38,9 +43,30 @@ module.exports = merge(common, {
         },
       ],
     }),
-    new MiniCssExtractPlugin({
-      filename: "[name].css",
-      path: path.resolve(__dirname, "dist/css"),
+    new CopyPlugin({
+      patterns: [
+        {
+          from: "./src/manifest.json",
+          to: "./manifest.json",
+        },
+      ],
     }),
+
+    new MiniCssExtractPlugin({ filename: "[name].[fullhash].css" }),
+    new HtmlWebpackPlugin({
+      template: "./src/index.html",
+      chunks: ["index"],
+      inject: true,
+      filename: "index.html",
+      minify: "auto",
+    }),
+    new HtmlWebpackPlugin({
+      template: "./src/popup.html",
+      chunks: ["popup"],
+      filename: "popup.html",
+      inject: true,
+      minify: "auto",
+    }),
+    new CleanWebpackPlugin(),
   ],
 });
